pass uint64 instead of float64 to store() and obviate pointless casts to float64 when uint64s are used internally anyway

+++ percent.go
@@ -5,6 +5,6 @@ import "sync/atomic"
 // (no-op) tracker for cases where status strings are not needed
 type percentTracker struct { current atomic.Uint64 }
 
-func (p *percentTracker) store(n float64, _ string) { p.current.Store(uint64(n)) }
-func (p *percentTracker) load()  any                { return p.current.Load()    }
-func (p *percentTracker) value(_ any)       string  { return ""                  }
+func (p *percentTracker) store(n uint64, _ string) { p.current.Store(n)      }
+func (p *percentTracker) load()  any               { return p.current.Load() }
+func (p *percentTracker) value(_ any)      string  { return ""               }
